Move cpu-cycles to metrics
