+++ common/mixnode-common/Cargo.toml
@@ -37,9 +37,3 @@ nym-task = { path = "../task" }
 nym-validator-client = { path = "../client-libs/validator-client" }
 nym-bin-common = { path = "../bin-common" }
 nym-metrics = { path = "../nym-metrics" }
-
-cfg-if = "1.0.0"
-cpu-cycles = { path = "../../cpu-cycles", optional = true }
-
-[features]
-cpucycles = ["cpu-cycles"]
