+++ common/nym-metrics/Cargo.toml
@@ -16,4 +16,11 @@ log = { workspace = true }
 dashmap = { workspace = true }
 regex = "1.1"
 lazy_static = "1.4"
+cfg-if = "1.0"
+
+cpu-cycles = { path = "../../cpu-cycles", optional = true }
+
+[features]
+cpucycles = ["cpu-cycles"]
+
